requirements.txt
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
PyJWT==1.7.1
SQLAlchemy==1.3.13
